docs/en-us/edx_val_api/requirements.txt
# Packages we need in order to build the docs, separated out so that rtfd.org
# can install them.
path.py
sphinx
sphinxcontrib-napoleon
markupsafe

# need to install requirements for application, in order to import Python code
-r ../../../requirements.txt
